Move shared tests for read_layer to shared module
